server: moved lnurl utils
